new folder for plotting tools and include all models to one single file instead of multiple files with the objective to simplify the project directory
