+++ plotting_tools/plot_avg_val_loss_and_rroc.py
