--- spec-file.txt
@@ -1,68 +1,0 @@
-# This file may be used to create an environment using:
-# $ conda create --name <env> --file <this file>
-# platform: linux-64
-@EXPLICIT
-https://conda.anaconda.org/anaconda/linux-64/_pytorch_select-0.2-gpu_0.tar.bz2
-https://conda.anaconda.org/anaconda/linux-64/blas-1.0-mkl.tar.bz2
-https://repo.anaconda.com/pkgs/main/linux-64/ca-certificates-2019.10.16-0.tar.bz2
-https://conda.anaconda.org/anaconda/linux-64/cudatoolkit-10.0.130-0.tar.bz2
-https://conda.anaconda.org/anaconda/linux-64/intel-openmp-2019.5-281.tar.bz2
-https://conda.anaconda.org/anaconda/linux-64/libgcc-ng-9.1.0-hdf63c60_0.tar.bz2
-https://conda.anaconda.org/anaconda/linux-64/libgfortran-ng-7.3.0-hdf63c60_0.tar.bz2
-https://conda.anaconda.org/anaconda/linux-64/libstdcxx-ng-9.1.0-hdf63c60_0.tar.bz2
-https://conda.anaconda.org/anaconda/linux-64/cudnn-7.3.1-cuda10.0_0.tar.bz2
-https://repo.anaconda.com/pkgs/main/linux-64/expat-2.2.6-he6710b0_0.conda
-https://repo.anaconda.com/pkgs/main/linux-64/icu-58.2-h9c2bf20_1.conda
-https://repo.anaconda.com/pkgs/main/linux-64/jpeg-9b-h024ee3a_2.conda
-https://conda.anaconda.org/anaconda/linux-64/libffi-3.2.1-h4deb6c0_3.tar.bz2
-https://repo.anaconda.com/pkgs/main/linux-64/libuuid-1.0.3-h1bed415_2.conda
-https://repo.anaconda.com/pkgs/main/linux-64/libxcb-1.13-h1bed415_1.conda
-https://conda.anaconda.org/anaconda/linux-64/mkl-2019.5-281.tar.bz2
-https://conda.anaconda.org/anaconda/linux-64/ncurses-6.1-he6710b0_1.tar.bz2
-https://repo.anaconda.com/pkgs/main/linux-64/openssl-1.1.1d-h7b6447c_3.tar.bz2
-https://repo.anaconda.com/pkgs/main/linux-64/pcre-8.43-he6710b0_0.conda
-https://conda.anaconda.org/anaconda/linux-64/xz-5.2.4-h14c3975_4.tar.bz2
-https://conda.anaconda.org/anaconda/linux-64/zlib-1.2.11-h7b6447c_3.tar.bz2
-https://repo.anaconda.com/pkgs/main/linux-64/glib-2.63.1-h5a9c865_0.tar.bz2
-https://conda.anaconda.org/anaconda/linux-64/libedit-3.1.20181209-hc058e9b_0.tar.bz2
-https://repo.anaconda.com/pkgs/main/linux-64/libpng-1.6.37-hbc83047_0.conda
-https://repo.anaconda.com/pkgs/main/linux-64/libxml2-2.9.9-hea5a465_1.conda
-https://conda.anaconda.org/anaconda/linux-64/readline-7.0-h7b6447c_5.tar.bz2
-https://conda.anaconda.org/anaconda/linux-64/tk-8.6.8-hbc83047_0.tar.bz2
-https://repo.anaconda.com/pkgs/main/linux-64/zstd-1.3.7-h0b5b093_0.conda
-https://repo.anaconda.com/pkgs/main/linux-64/dbus-1.13.12-h746ee38_0.tar.bz2
-https://repo.anaconda.com/pkgs/main/linux-64/freetype-2.9.1-h8a8886c_1.conda
-https://repo.anaconda.com/pkgs/main/linux-64/gstreamer-1.14.0-hb453b48_1.conda
-https://repo.anaconda.com/pkgs/main/linux-64/libtiff-4.1.0-h2733197_0.tar.bz2
-https://conda.anaconda.org/anaconda/linux-64/sqlite-3.30.1-h7b6447c_0.tar.bz2
-https://repo.anaconda.com/pkgs/main/linux-64/fontconfig-2.13.0-h9420a91_0.conda
-https://repo.anaconda.com/pkgs/main/linux-64/gst-plugins-base-1.14.0-hbbd80ab_1.conda
-https://conda.anaconda.org/anaconda/linux-64/python-3.7.5-h0371630_0.tar.bz2
-https://repo.anaconda.com/pkgs/main/linux-64/certifi-2019.9.11-py37_0.tar.bz2
-https://repo.anaconda.com/pkgs/main/linux-64/kiwisolver-1.1.0-py37he6710b0_0.conda
-https://conda.anaconda.org/anaconda/linux-64/ninja-1.9.0-py37hfd86e86_0.tar.bz2
-https://repo.anaconda.com/pkgs/main/linux-64/olefile-0.46-py37_0.conda
-https://conda.anaconda.org/anaconda/linux-64/pycparser-2.19-py37_0.tar.bz2
-https://repo.anaconda.com/pkgs/main/noarch/pyparsing-2.4.4-py_0.tar.bz2
-https://repo.anaconda.com/pkgs/main/noarch/pytz-2019.3-py_0.tar.bz2
-https://repo.anaconda.com/pkgs/main/linux-64/qt-5.9.7-h5867ecd_1.conda
-https://repo.anaconda.com/pkgs/main/linux-64/sip-4.19.8-py37hf484d3e_0.conda
-https://conda.anaconda.org/anaconda/linux-64/six-1.13.0-py37_0.tar.bz2
-https://repo.anaconda.com/pkgs/main/linux-64/tornado-6.0.3-py37h7b6447c_0.conda
-https://conda.anaconda.org/anaconda/linux-64/cffi-1.13.1-py37h2e261b9_0.tar.bz2
-https://repo.anaconda.com/pkgs/main/linux-64/cycler-0.10.0-py37_0.conda
-https://conda.anaconda.org/anaconda/linux-64/mkl-service-2.3.0-py37he904b0f_0.tar.bz2
-https://repo.anaconda.com/pkgs/main/linux-64/pillow-6.2.1-py37h34e0f95_0.tar.bz2
-https://repo.anaconda.com/pkgs/main/linux-64/pyqt-5.9.2-py37h05f1152_2.conda
-https://repo.anaconda.com/pkgs/main/noarch/python-dateutil-2.8.1-py_0.tar.bz2
-https://conda.anaconda.org/anaconda/linux-64/setuptools-41.6.0-py37_0.tar.bz2
-https://conda.anaconda.org/anaconda/linux-64/numpy-base-1.17.3-py37hde5b4d6_0.tar.bz2
-https://conda.anaconda.org/anaconda/linux-64/wheel-0.33.6-py37_0.tar.bz2
-https://conda.anaconda.org/anaconda/linux-64/pip-19.3.1-py37_0.tar.bz2
-https://conda.anaconda.org/anaconda/linux-64/mkl_fft-1.0.15-py37ha843d7b_0.tar.bz2
-https://conda.anaconda.org/anaconda/linux-64/mkl_random-1.1.0-py37hd6b4f25_0.tar.bz2
-https://conda.anaconda.org/anaconda/linux-64/numpy-1.17.3-py37hd14ec0e_0.tar.bz2
-https://repo.anaconda.com/pkgs/main/linux-64/matplotlib-3.1.1-py37h5429711_0.conda
-https://conda.anaconda.org/anaconda/linux-64/pytorch-1.2.0-cuda100py37h938c94c_0.tar.bz2
-https://conda.anaconda.org/anaconda/linux-64/pytorch-gpu-1.2.0-0.tar.bz2
-https://repo.anaconda.com/pkgs/main/linux-64/torchvision-0.4.0-cuda100py37hecfc37a_0.conda
